requirements.txt
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
python-dotenv==0.21.0
pytz==2024.1